+++ src/components/PDFExport/index.tsx
@@ -3,6 +3,8 @@ import axios from 'axios';
 import { Base64 } from 'js-base64';
 import escape from 'lodash';
 
+import { downloadBlob } from 'utils/downloadFile';
+
 import downloadSVG from './download.svg';
 
 type PDFExportProps = {
@@ -11,26 +13,6 @@ type PDFExportProps = {
   children: React.ReactNode;
   label?: string;
 };
-
-function downloadBlob(filename: string, blob: Blob) {
-  // This approach to downloading files works fine in the tests I've done in Chrome
-  // with PDF files that are < 100kB. For larger files we might need to
-  // instead redirect the browser to a URL that returns the file. That
-  // approach is complicated by using JWTs for auth.
-  //
-  // TODO test this in various browsers. Some reports say this might not work
-  // properly in Firefox and that firing a MouseEvent is required instead.
-  const blobUrl = URL.createObjectURL(blob);
-
-  const link = document.createElement('a');
-  link.href = blobUrl;
-  link.download = filename;
-
-  // This downloads the file to the user's machine.
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-}
 
 function generatePDF(filename: string, content: string) {
   axios
