+++ src/views/Accessibility/AccessibilityRequest/List/index.tsx
@@ -11,6 +11,7 @@ import { NavLink, SecondaryNav } from 'components/shared/SecondaryNav';
 import useMessage from 'hooks/useMessage';
 import GetAccessibilityRequestsQuery from 'queries/GetAccessibilityRequestsQuery';
 import { GetAccessibilityRequests } from 'queries/types/GetAccessibilityRequests';
+import { downloadBlob } from 'utils/downloadFile';
 
 const List = () => {
   const { t } = useTranslation('home');
@@ -40,26 +41,6 @@ const List = () => {
     data.accessibilityRequests.edges.map(edge => {
       return edge.node;
     });
-
-  function downloadBlob(filename: string, blob: Blob) {
-    // This approach to downloading files works fine in the tests I've done in Chrome
-    // with PDF files that are < 100kB. For larger files we might need to
-    // instead redirect the browser to a URL that returns the file. That
-    // approach is complicated by using JWTs for auth.
-    //
-    // TODO test this in various browsers. Some reports say this might not work
-    // properly in Firefox and that firing a MouseEvent is required instead.
-    const blobUrl = URL.createObjectURL(blob);
-
-    const link = document.createElement('a');
-    link.href = blobUrl;
-    link.download = filename;
-
-    // This downloads the file to the user's machine.
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  }
 
   function fetchCSV() {
     axios
